requirements.txt
django==2.2.2
django-filter==2.2.0
django-filters==0.2.1
django-dynamic-formsets==0.0.8
psycopg2-binary
django-leaflet
bcrypt
django[argon2]
pillow
django_jalali
jdatetime
kavenegar
